Raise revision to 1.2.1 (#30)
* Raise revision to 1.2.1

Raise revision number to 1.2.1 in files that specify revision
in preparation for next release due to license change.
The files in question are the C-API makefile, readme and the get revision function
as well as the ascii doc specification.
Static revision numbers in specification are replaced by revnumber variable
for ease of use and change.
Additionally updated the authors in notice file because dates have changed 
and contributors were added.

* [doc] Add RELEASENOTES and update Release Notes within c-api readme

Signed-off-by: Sebastian_3DMS <sebastian.tuttas@3d-mapping.de>

* Update NOTICE

Update contributors in NOTICE file in preparation of next release.
NOTICE file is still subject to change.

Signed-off-by: Patrick Kuemmerle patrick.kuemmerle@3d-mapping.de

* Update NOTICE

merged copyright statement of Daimler AG - Jochen Rauh

* [doc] update releasenotes for upcoming version

---------

Signed-off-by: Sebastian_3DMS <sebastian.tuttas@3d-mapping.de>
Signed-off-by: Patrick Kuemmerle patrick.kuemmerle@3d-mapping.de
Co-authored-by: Sebastian_3DMS <sebastian.tuttas@3d-mapping.de>
Co-authored-by: Jochen Rauh <47066552+jorauh@users.noreply.github.com>